PlaySoundメソッドをMusicTableコンポーネント内に移動
